Remove font awesome config
